Address several inconsistencies in the built-in types.
The class hierarchy for the immediate types that overflow to a heap
type was wrong.  Previously, the immediate type was erroneously a
subtype of the heap type.  This change introduces abstract superclasses
that the immediate types and heap types now derive from.

The BoundMethod and Set types were missing from the macro list of
built-in types and were being ignored by tests.  This list has been
updated.  The macro list and related lists of types have been sorted
to make future omissions more prominent.

The name of the bytearray type has been corrected to match CPython.

Based on Facebook D7432684