Remove out-of-memory assertions.
newFunction had incorrectly checked the return value of
heap->createFunction() for nullptr to detect an out-of-memory error.
However, when an allocation fails, an Error object is returned, not
nullptr.

Nevertheless, an out-of-memory error will cause the allocator to
fail fast so there is no way for an Error object to be propagated up
to the constructor routines.

Based on Facebook D9817463