Add layoutIdName() in runtime-test.cpp
...and use it for a better error message in
BuiltinClassIdsTest.HasClassObject. I added empty classes for Gen and Coro, and
this also exposed a bunch of types that were missing from the
INTRINSIC_*_CLASS_NAMES macros in objects.h. The switch statement in
layoutIdName() should prevent any more omissions from these macros in the
future.

This is intentionally a low-level function with no external dependencies, so it
only supports builtin LayoutIds.

Based on Facebook D9763970