Enforce a coding style
This diff adds a `.clang-format` file taken from the HHVM project,
a tool to auto-format the entire codebase, and lint integration
with `arcanist`.

The script `util/format.sh`, taken from the Hermes project, will
automatically apply the style specified in `.clang-format` to
the appropriate files in the codebase. By default, it will only
attempt to format modififed files or, if there are none, files
that were touched in the last commit. If supplied with the `-a`
argument, it will format all `.cpp` and `.h` files not in the
`third-party` directory.

Additionally, a new arcanist linter will report a warning if
any modified files to do not conform to the style outlined in
`.clang-format`. Note that it does not automatically reformat the
file on behalf of the user. This is possible, but I don't feel like
writing a diff parser in PHP right now.

Finally, `util/format.sh -a` was used to format the codebase.

Based on Facebook D5959863