Add the _stat module and its API dependencies
This change adds PyModule_AddIntConstant which _stat uses to add
integer constants to its module.  It also rearranges some code in
the module object unit test to make the module definition private
to each test.  To test this, the module definition is now tracked
in the module object when a module is initialized from a PyModuleDef.

Based on Facebook D8960191