Add support for deleting attributes from a layout
Title sums it up. Note that in the current implementation deleting
an in-object attribute will leave a hole in the instance (a word
in the instance will no longer be usable). Re-adding the attribute
will add it to the overflow array. If this proves to be a problem
we can support re-using slots in the instance in the future.

Based on Facebook D7352883