Make the SymbolId type useful outside of the Symbols class
The objective of this change is to make SymbolId a useful substitute
for an Object* to an interned string in contexts where an predefined
interned string is used.

There are two intended effects.  First, handles are no longer needed
to reference to a predefined interned string.  Second, references to
predefined interned strings can now be stored in static structures as
their SymbolId is an ordinary small integer.

Based on Facebook D8718649