Revert "Add phabricator code coverage support"
This change uses an environment variable to configure coverge.
However, the convention in CMake is either

* To add a CMake build type for a coverage mode, just as one has
build types for release and debug mode.

* To create a CMake option that turns coverage support on or off.

Whether we use the first or second strategy, this allows coverage to
be explicitly specified on the CMake invocation that creates build
files.  Following the conventions above, this would be something like

* `cmake -DCMAKE_BUILD_TYPE=Coverage`

* `cmake -DCOVERAGE=ON`

This command failed right away on my system which is configured to use
clang

```
$ arc unit
...
Scanning dependencies of target python
[ 54%] Building CXX object CMakeFiles/python.dir/main.cpp.o
[ 57%] Linking CXX executable python
ld: library not found for -lgcov
clang: error: linker command failed with exit code 1 (use -v to see invocation)
make[2]: *** [python] Error 1
make[1]: *** [CMakeFiles/python.dir/all] Error 2
make[1]: *** Waiting for unfinished jobs....
[ 60%] Linking CXX static library libgmock.a
[ 66%] Linking CXX static library libgtest.a
[ 66%] Linking CXX static library libgmock_main.a
[ 66%] Built target gmock
[ 66%] Built target gtest
[ 66%] Built target gmock_main
make: *** [all] Error 2
   FAIL  Exit status non-zero (2): PYRO_ENABLE_COVERAGE=1 "/Users/cshapiro/CLionProjects/pyro/arcanist/unit/run.sh" "/Users/cshapiro/CLionProjects/pyro" "/var/folders/rq/t98j9b5n799b691x2n2ky_6d2z9525/T/tmp.6eY4Pc2z" "/var/folders/rq/t98j9b5n799b691x2n2ky_6d2z9525/T/tmp.6eY4Pc2z/test_output.xml"
```

We should support both GCC and LLVM.  I think the flags should be the
same for both compilers (and their linkers) so we may just be asking
for coverage support in an overly restrictly way.

Coverage seems to be enabled by default with `arc unit` rather than
through an explicit invocation of `arc unit --coverage`.  I am not
sure what the defaults should be and it is not obvious to from the
`arc unit` options.

Based on Facebook D6150929