[Pyro] Tests: Cleanup binary
This cleans up the tests and benchmark binaries and combines them into one based on: https://github.com/google/benchmark/issues/145.

This makes it so that in one binary by default, we run all the unit tests. Then, if specified, the benchmarks are run without the unit tests.

Based on Facebook D7406247