Transition layouts when adding in-object attributes
This change decouples instance size calcuation from the logic to
compute the set of in-object attributes. Previously, we
incorrectly treated all in-object attributes as being set in the
constructor. As a result, the runtime might crash when it encountered
code like the following:

  class Foo:
    def __init__(self, x):
      self.foo = 'foo'
      if x:
        self.bar = 'bar'

  print(Foo(False).bar)

The old code would calculate an instance size of 2 for Foo and treat
`foo` and `bar` as in-object attributes. Since `bar` is conditionally
initialized based on the value of `x`, we would incorrectly assume
that `bar` was present when `x` is `False`.

This change computes the instance size using the original logic, but
dynamically fills in the set of in-object attributes as they are
initialized.

Based on Facebook D7457738