Implement hashing for all instance types.
Make use of the header hash code field.

Heap objects have few bits reserved in their header for an identity
hash code.  This value is initially zero indicating that the object
has not been hashed.  When a hash code is requested, a hash code is
computed and inserted into the header.  Hash codes for strings and
bytes are derived from their contents.  Other types use bits from a
random number generator.

The runtime now reads a few words from the system entropy source in
order to seed the random state and produce a secret key to randomize
the hash code for strings and bytes.

Based on Facebook D6190882