Make missing roots known to the garbage collector.
This change extends the root visitor to visit the thread stack and
makes the raw pointer arguments of two to constructors known to the
visitor by putting them into handles.  This also disabiguates some
names of local variables that share the same name as variables in
an outer scope.

Based on Facebook D7061063