Finish implementation of tuple.__new__
One of the arguments to `tuple.__new__` is an iterable. Implement the code that takes the iterable and makes a tuple from it.
Since tuples are immutable and fixed-length, there are some hoops to jump through. Use a doubling tuple as storage, and add a special function `at_put` which takes a tuple, index, value and inserts the value at that index in the tuple.

Based on Facebook D9504697