Fixed two bugs in SetIterator
`SetIterator::kIndexOffset` and `SetIterator::kConsumedCountOffset`
were referring to the same offset, this diff fixes that.
Also, the indexing in `SetIterator::next()` was flawed,
since the index into the set's bucket array needs to be incremented
by two and not one. Extended the existing `SetIterator` test case
to test for iterator exhaustion. Also, added a new test case
to test iterating over an empty set.

Based on Facebook D10054514