Implement builtin issubclass
The builtin `issubclass(T, X)` checks if `T` is a subclass of `X`. This works on the type level, as opposed to `isinstance` which works on instances of types.
If `X` is a tuple, then if `T` is a subclass of any member of `X`, it returns true.
It throws if any element in `X` is not a type.

Based on Facebook D9386059