[Pyro] Introduce ApiTypeHandle
This introduces the ApiTypeHandle into the runtime.

ApiTypeHandle subclasses from PyTypeObject in order to retain it's member memory layout. Furthermore, this adds some extra functionality such as adding a extra flag slot in order to distinguish the builtin extension types that are part of the runtime.

Finally, this adds a static builder which help the extension types get initialized.

Based on Facebook D9016200