[pyro]: Reduced the number of allocations in ListBuiltins::dunderAdd.
Instead of using `listExtend` to copy two lists in `ListBuiltins::dunderAdd`,
create an ObjectArray with enough capacity to fit elements from both lists
and copy elements from both lists into it and set this array as the new list's
backing array. Added a `ObjectArray::copyToWithOffset` method to help with this.

Based on Facebook D9260540