Reconcile metaclasses and the type system
This change reconciles metaclasses with the runtime's type system.
Classes are now created with an explicitly supplied metaclass layout id.

Some things to note:

`Object::isClass()` will return `true` if the receiver is an instance
of `Class` whose metaclass is `type`. In this way, `isClass` has the same
semantics as CPython's `PyType_CheckExact`. For example, `isClass()` would
return true for all of the classes in the following Python code.

```
class Foo(type):
  pass

class Bar(Foo):
  pass
```

If this seems counterintuitive, consider that the above is equivalent
to the following code:

```
class Foo(type, metaclass=type):
  pass

class Bar(Foo, metaclass=type):
  pass
```

which, in turn, is equivalent to:

```
Foo = type('Foo', (type,), {})
Bar = type('Bar', (Foo,), {})
```

When viewed this way it's hopefully more obvious that `Foo` and `Bar`
are in fact concrete instances of `type`.

In contrast, `Object::isClass()` will return `false` for `Baz` in the
following snippet:

```
class Baz(metaclass=Foo):
  pass
```

Again, if that seems counterintuitive, recall that the code is equivalent to

```
Baz = Foo('Baz', (), {})
```

Hopefully that syntatic form makes it clearer that `Baz` is an instance of
`Foo`, rather than `type`.

I've added the `isInstanceOfClass` predicate to `Runtime` that provides a
way to check if an object is an instance of `type` or a metaclass that
inherits from `type`. This is equivalent to `PyType_Check` in CPython. We
should take care to use `isInstanceOfClass` in runtime code where any
instance of a subclass of `type` will suffice (likely most cases).

Based on Facebook D8952728