[Pyro] Add TrackedAllocations linked list
This adds a linked list which tracks all the allocated extension objects.

It does this by mallocing an extra two pointers which have a point to the previous and the next allocated object. Thus, when it's time to free all the memory, it's easy to walk through the list and clear all the memory.

Based on Facebook D8415061