clang-tidy: Allow 1-line if statements without braces
This will allow if statements on a single line without any braces when
the whole statement (condition and body) fits within the line length limit. I
think this is much more readable for short if statements, especially if there
are more than one of them in a row.

This is allowed:
```
if (someThing()) return false;
```

This is not:
```
if (someThing())
  return false;
```

For the latter case, when the whole thing fits on one line, clang-format will
turn it into the former. When it doesn't fit on one line, clang-format will
spread it across as many lines as needed and then clang-tidy will yell at you
to add braces.

Based on Facebook D10047807