Implemented set.intersection() and the & and &= operator on sets
Summary: Added the `set.intersect()` builtin method and support for `&` (`__and__`)  and `&=` (`__iand__`) operators on Sets. All three of these methods are backed by a new runtime method `Runtime::setIntersection()`. `Interpreter::doBinaryAnd()` was specialized only for `SmallInt`s, made it generic and dispatch to the receiver's `__and__` method. Added `SmallIntBuiltins::dunderAnd()`, as a side effect of making `Interpreter::doBinaryAnd()` generic. `set.intersect()` with no arguments returns a copy of the set; added `Runtime::setCopy()` and an overloaded `Runtime::newSet(word)` to support this. Added a `getPythonIterator()` helper function in runtime-tests to reduce the duplication of an oft  repeated python snippet.

Based on Facebook D10158068