Implement PyUnicode_AsUTF8(), fix PyUnicode_AsUTF8AndSize()
PyUnicode_AsUTF8(obj) is shorthand for PyUnicode_AsUTF8AndSize(obj,
nullptr). The bigger issue is that both functions are supposed to return a
cached char* that's owned by the PyObject* it came from. Add a basic cache to
ApiHandle to support this and fix up a few callsites that were incorrectly
freeing data that didn't belong to them.

Based on Facebook D10365833