Implement PyObject_Generic{Get,Set}Attr
Implement the generic getters and setters for object attributes. This API should be used to directly set and get elements from a Type, Module, and Instance.

This also refactors the runtime to do the String checks at the beginning of attributeAt and attributeAtPut.

Finally, a new helper function was added to thread.h to quickly check if there is a pending exception as a bool rather than checking for the pending exceptions type.

Based on Facebook D9729597