Fix Runtime::importModule for extension modules
`PyModule_Create` should only create a module object; it should not call `runtime->addModule`. If we remove that line, it reveals a problem in `Runtime::importModule` - if it finds the module name in inittab, it needs to call `addModule` directly.

Fix `Runtime::importModule` and remove the line from `PyModule_Create` - this also requires minor changes in some test to make them pass.

Based on Facebook D10367997