[pyro]: Added a set_iterator builtin type
Added a `SetIterator` builtin.
Added the `__iter__` method to `Set`s which returns a `SetIterator` instance.
Added `__iter__`, `__next__` and `__length_hint__` methods to the `SetIterator` class.
Unlike the previously added iterators for `List`, `Range` and `Tuple` objects,
we need to keep track of the number of consumed elements in the iterators.
This is to help with the `__length_hint__` method.

Based on Facebook D9555535