Replace ints with strings in tests checking refcount
These tests are failing with CPython because it has static PyObjects for very small numbers (a few hundred), so ref count is not as we expect. We will want to implement this optimization as well, so this might be a problem for us in the future as well.

The type doesn't matter in those cases, let's make them strings instead

Based on Facebook D10842338