Implement str.split
`str.split` takes a separator, and returns a list of strings that are in between the separator. It puts an empty string if two separators are next to each other. If `maxsplit` is specified, it only goes up to that number of splits, and then returns the rest of the string unmodified.

NOTE: Since `importlib` does not use the case of `split` with no separator, that is left unimplemented for now.

Based on Facebook D9467953