Move escaped newlines to the right to reduce diff sizes
Previously, escaped newlines were aligned vertically in the leftmost
column as determined by the longest line in the containing macro.
This had an undesirable effect of causing all of the escaped newlines
in a macro to move when the length of the longest line in a macro
changed.  For large macros like the bytecode list, this could inflate
the diff size considerably and add noise to the commit history.

This change switches to placing escaped newlines in the rightmost
column where they are generally immune to changes in the longest line
of a macro.

Based on Facebook D8643781