Eliminate code duplication between Runtime::dictUpdate and Runtime::dictMerge
`dictUpdate` and `dictMerge` shared the same code,
except that `dictMerge` also has additional checks to
ensure that all keys are strings and that there are no repeated keys.
Refactored `dictMerge` into a private templated method, that performs
the additional checks if the template parameter `merge` is true.
The public methods `dictUpdate` and `dictMerge` both call the aforementioned private method.

Based on Facebook D9749314