Change format.sh to format.py for greater safety
Bash scripts suck and have platform-specific behavior.
Change this to be a python file to enable cross-platform formatting ("all" was broken on mac due to broken `find`).

This also makes it much easier to extend it in the future.

It ignores changes to files in `third-party`, `benchmarks`, and `ext`

Based on Facebook D9038815