Make Handle<T> inherit from T
This is kind of gross, but it will allow us to keep doing things like
`SmallInt::fromWord()` between D10111147 and when we move the static functions
to the new handle classes.

The only practical downside I can think of is that Handle<T> will implicitly
convert to T now, which might let some GC-safety bugs creep in while we're in
this state.

Based on Facebook D10184841