Introduce testing::moduleSet for C-API testing
Some of our CAPI tests currently rely on incorrect behavior of `PyModule_Create`. To fix this, we either need to implement and support `PyImport_AppendInittab`, which is problematic (it requires a static array of modules, which means tests are affecting each other if they use it :-/) or avoid importing custom modules in tests.

This way we can avoid importing custom modules via inittab - instead of
```
PyRun_SimpleString(R"(
import mymodule
// some code with mymodule
)")
```

we can now do

```
testing::moduleSet("__main__", "mymodule", module_obj);
PyRun_SimpleString(R"(
// some code with mymodule
)")
```
This removes the dependency on import, which is not what those tests should be testing.

Based on Facebook D10409811