[pyro]: General cleanups in list unpack, tuple unpack, listExtend
Renamed the `iterator` parameter in `Runtime::listExtend` to `iterable`.
Scope handles in `Interpreter::doBuildListUnpack` and `Interpreter::doBuildTupleUnpack`
to the interpreter thread, removed the use of nested scopes in the aforementioned functions.
Unroll loops with the `EXPECT_*` macros in tests for list and tuple unpacking.
This diff was split from the larger diff D8323114.

Based on Facebook D8351273