Move the attribute tests next to the rest of the runtime tests.
The general rule is to put test for implementation code in
<filename>.cpp into <filename>-test.cpp.  The only exception to
this rule in our codebase was attribute-test.cpp.  That code was
supposed to be moved into runtime-attribute-test.cpp after the
attribute-related code in runtime.cpp was itself moved into
runtime-attribute.cpp.  However, nobody has gotten around to that
and this exception to the rule has remained.

By moving this code, our test code is more consistently located and
more discoverable.  Furthermore, the rule about where to put your
code is actually reflected in our source code.

Based on Facebook D8411863