[Pyro] Extensions: Partial implemnentation of PyUnicode_AsUT8AndSize
Add in a partial implementation of PyUnicode_AsUT8AndSize. This is still missing the caching functionality. Meaning that all generated CStrings from this function will be newly allocated.

Currently, this function is required to be able to link in some of the unmodified functions required for PyType_Ready. None of them require the identity functionality so it's fine to push this feature to be implemented later.

This change also adds a modification to the String::toCString() api, adding a length component to only set the specified number of chars in the newly allocated CString. This defaults to the full string.

Based on Facebook D8127516