Extend keyword calling to support arbitrary callable objects
This adds support to the interpreter to perform keyword calls for arbitrary callable
objects. The first use case of this is constructing `ImportError`, which allows keyword
arguments (e.g. `ImportError(name='foo')`). The logic for canonicalizing a stack frame
that is needed by both `Interpreter::call` and `Interpreter::callKw` is extracted into
a set of helpers.

Based on Facebook D9328898