[Pyro] Enum class LayoutId
Created a scoped enum for LayoutId with an underlying type of word.

The layoutId names were causing conflicts with other enums in the runtime. I.e. ExtensionTypes::kType. By creating an enum class, it will scope all the names. Plus, it makes the use of the enums to be strictly typed making it more type safe.

Furthermore, this makes printing this in the debugger much better.

Based on Facebook D8383713