[Pyro] Enable target determinator
This adds target determinator into the pyro repo.

Target Determinator will run as a first step and will determine the different builds that it needs to perform to verify the working builds on the different subsystems.

Separately, a set of unit is run to verify their correctness.

Based on Facebook D8832262