[Pyro][noop] Add helper macros for ext functions
This adds to macros to our copy of CPython's object.h.

Py_AsPyObject: Casts an ApiHandle to a PyObject
Py_AsApiHandle: Casts a PyObject to an ApiHandle.

These are two helper functions to ease up the conversion between ApiHandles and PyObjects.

Note that ApiHandle is isomorphic to a PyObject, thus, casting between them is valid.

Based on Facebook D8074317