Specialize Runtime::setLookup for inserts and lookups
Sets in Pyro are hashtables with linear probing for collision resolution.
The idea of tombstone slots in the hash table is to continue
the lookup if a tombstone record is found and to end it if an empty
slot is found, but this isn't needed while looking up a bucket
for insertion. For insertion,it's fine to overwrite a tombstone
record. This change makes `setLookup` exit early for the insert case.

Based on Facebook D9991301