Move object initialization out of the heap for GC safety.
Previously, many of the methods for low-level object allocation
would accept some arguments in the form of raw pointers that would
be used to initialize a newly allocated object.  However, if the
allocator ever retried an allocation, these pointers would become
dangling.

This change defaults to None initializing object attributes and
pushes almost all of the initialization into the constructor where
the result objects can be stored in handles.  A few classes had to
assumed zero initialization and there constructors were updated.

Based on Facebook D7436377