Re-initialize PyModuleDefs on every test run
Marking PyModuleDefs as `static` solves the problems described on D10438198, but this could have a surprising behavior if we run a test repeatedly. We can't guarantee that the struct is not modified during the test run, so we should re-initialize it every time.

This is formally a "copy-list-initialization", which does "aggregate initialization", which does "value initialization" for the remaining members, which is "zero-initialization" for non-structs.

Based on Facebook D10500911