[pyro][easy]: Minor cleanups
Fixes from review feedback from D8510628, D8515035 and D8515495,
which came after the stacked diff landed.
Changed the error message on TypeError when the self is not of the right type in `__add__` and `__sub__`
methods for `SmallInteger`, `Double` builtins, and `__add__` method for the `List` builtin.
Added the `Integer::doubleValue()` method to get the double representation of `SmallIntegers` and `LargeInteger` objects.
Used the `Integer::doubleValue()` method to simplify `builtinDoubleAdd` and `builtinDoubleSub`.
Broke `IntBuiltinsTest_InplaceOps` into multiple test cases.

Based on Facebook D8542999