Use numeric symbol ID instead of raw Object pointers to symbols
Several methods used during the bootstrapping of the runtime would
accept a raw Object* parameter as a convenience for accepting a
pointer to a symbol.  This required the callee to immediate store
the pointer into a handle or carefully avoid any operations that
could allocate.

This change switches to the much safer idiom of accepting a symbol
ID which is a compile time constant making bootstrap module operations
much simpler and safer.

Based on Facebook D8830159