Util: Improve source generation speed
Summary: In preparation to compiling in CPython *.cpp files, This change refactors the generate_cpython_source.py script to only run the regexes on their respective *.h or *.cpp file. Plus, this will also just do a file copy if a source is not actually modified.

Based on Facebook D9724890