Break-up the exception state
Previously, when raising an exception, a reified exception object
was stored in the thread object.  This does not match the behavior
of the Python API which expects to keep the exception type, and
optional value, and a traceback object in the thread state.

This models the current exception state as three separate quantities.
It also makes the raising interfaces record the type of the pending
exception in addition to a value.

This also renames the methods for raising exceptions to use the
Python terminology of "raise" instead of "throw".  It also uses the
preposition "with" to more accurately reflect that the value is a
component of the exception state, not what the exception state is
being created from.

Based on Facebook D9994416