[Pyro] C-API: Implmement setting and checking exceptions
This partially implements some exception setting and checking in the C-API. This will be used extensively in the C-API tests.

This is partially implemented because exception types can not be instantiated yet in the runtime. Therefore, PyErr_SetString currently defaults to SystemError. Once that piece of runtime is implemented, it will allow PyErr_SetString to correctly set the exception based on the exception type.

Furthermore, this adds a non C-API function called: `_PyErr_ExceptionMessageMatches`. This will be used in C-API tests to test that a certain exception string was set.

Based on Facebook D9355057