Switch to a compact object header.
Previously, object headers pointed to the instance's class object and
the class object was used to size objects and to retrive the type.
This indirection adds runtime overhead and adds some complexity to the
garbage collector.

This change switches the header to be a bit field that contains a
count, a hash code, a class ID, and a format ID all packed into a
single 64-bit integer.  By colocating this information with the
instance, most metadata queries can be answered without indirection.

Most of these fields are logically smaller than what would otherwise
be available with an indirect representation, but they should be big
enough.  For example, the class ID is a 20-bit field, which is an
order of magnitude larger than the number of classes in our code base.

In addition, this change introduces some rationalization of the use of
integer types, switching to a word for integers that range over the
size of an object.  In other cases, more specific integer types were
used such as an int32 in favor on an int, in order to improve
readability.

Based on Facebook D6117740