Implement str.__repr__
`str.__repr__` prints out the string with wrapping quotes, and escapes non-printable characters.

If there are both types of quotes in the string, single quotes are escaped and single quotes wrap the result.
If there are only single quotes, then the wrappers are changed to double quotes.

NOTE: This currently copies all bytes beyond the non-printable ASCII range as-is.

Based on Facebook D9415871