[Pyro] Add newStringFromFormat
There's a lot of times when we need to create a python string based on some format and arguments (like `sprintf`).
Make a convenience function in `Runtime` to do this for us.

It's especially useful in exception messages to give some context.

Added a `FORMAT_ATTRIBUTE` macro that optionally expands to `__attribute__((format(...)))`, an attribute that allows the compiler to check a printf-like syntax for type errors at compile-time.

Based on Facebook D8919498