Always pass the result of a throw upward to the caller.
With this change the return value of an exception throwing subroutine
is returned upward to the caller instead of sometimes just throwing it
away.

In addition, the signature of the exception throwing subroutines now
accept an object instead of a string.  In fact, an exception message
is just the string representation of its value.  This is more correct
and removes some casts.

Based on Facebook D8520010