Remove &PyTuple_GET_ITEM from functools module
There's an optimization in `functools.partial` for cases when we don't need to concatenate lists of arguments (either functool.partial was created with no positional arguments or it's called without any positional arguments).

This uses `&PyTuple_GET_ITEM` and `_PyObject_FastCallDict` that we can't support because it relies on `PyObject**`.

Remove this optimization. I'll copy this onto cpython repo as well once accepted.

Based on Facebook D10844415