[Pyro] Refactor ApiHandle creation to the ApiHandle object
This moves the runtime's `asApiHandle` and `asBorrowedApiHandle` into the ApiHandle class. They were further renamed to: `fromObject` and `fromBorrowedObject` to keep consistency with the `fromPyObject` API.

The Runtime itself does not require to create ApiHandles, thus, these should be handled by the ApiHandle itself.

Note: This does not change any functionality it is a pure refactoring.

Based on Facebook D8810319