[pyro][easy]: tuple builtins refactor
Refactored tuple-builtins to turn all builtin functions into static methods on a class.
Now that it's namespaced, it makes sense to rename `tupleSlice` to `TupleBuiltins::slice`.
Made `TupleBuiltins::slice` private, and added manual manual expansions of the FRIEND_TEST
to the class body to keep `TupleBuiltins::slice` visible for the relevant tests.

Based on Facebook D9360047