Add Layouts (nee Hidden Classes) and overflow attributes
This diff introduces the Layout type, which describes the shape of
an instance, and splits instance attributes into two classes: in-object
attributes, which exist directly in the instance, and overflow attributes,
which are stored in an object array pointed to by the last word of the
instance. Graphically, this looks like:

  Instance                                   ObjectArray
  +---------------------------+     +------->+--------------------------+
  | First in-object attribute |     |        | First overflow attribute |
  +---------------------------+     |        +--------------------------+
  |            ...            |     |        |           ...            |
  +---------------------------+     |        +--------------------------+
  | Last in-object attribute  |     |        | Last overflow attribute  |
  +---------------------------+     |        +--------------------------+
  | Overflow Attributes       +-----+
  +---------------------------+

Each instance is associated with a layout (whose id is stored in the header
word, taking the place of the class id). The layout acts as a roadmap for
the instance; it describes where to find each attribute.

In general, instances of the same class will have the same shape. Idiomatic
Python typically initializes attributes in the same order for instances of
the same class. Ideally, we would be able to share the same concrete Layout
between two instances of the same shape. This both reduces memory overhead
and enables effective caching of attribute location.

To achieve structural sharing, layouts form an immutable DAG. Every class
has a root layout that contains only in-object attributes. When an instance
is created, it is assigned the root layout of its class. When a shape
altering mutation to the instance occurs (e.g. adding an attribute), a label
for the mutation (e.g. 'add:<attribute>') is created. If there is an edge
with the same label on the instance's layout, the edge is followed and the
instance is assigned the resulting layout. If there is no edge, a new layout
is created, an edge with the label is inserted between the two layouts, and
the instance is assigned the new layout.

Based on Facebook D7306783