Begin cleanup of Object/Handle C++ types
This is the first in a series of diffs to clean up our Object/Handle
system. The desired end state is roughly:

- All classes in the current `Object` type hierarchy will have `Raw` prepended
  to their names.
- The new `RawObject` types will be wrappers around a `uword` that we pass by
  value, rather than using pointers (see D9927060 for motivation).
- There will be a parallel hierarchy of classes using the old names from the
  `Object` hierarchy, which will take the place of `Handle<T>` (so you'll use
  `Object` instead of `Handle<Object>`, `Str` instead of `Handle<Str>`,
  etc...).
- All non-static member functions (mostly getters and setters) from the
  `RawObject` types will be moved to the new handle types, which should be the
  primary API people interact with in C++. This will make it much harder to
  write GC-unsafe code.

I've tried to split the plan up into a number of different pieces, each of
which should either be a) completely mechanical and done 100% by a script or b)
small enough that manually reviewing the whole thing is feasible.

== Step 1 (this diff)
Add `using RawFoo = struct Foo*;` for all `Object` types, and replace all
occurrences of `Foo*` with `RawFoo`. This preempts most of the boring changes
needed by in D9927060, which should make it easier to review.

== Step 2 (D9927060)
Change all `Object` types to be value types, and change the typedefs added by
step 1 to `using RawObject = Object;`

== Step 3
Remove all `operator->()` functions from `RawObject` types that were added in
step 2 to ease migration. This will probably be deferred until after steps 5+.

== Step 4 (D10184841, D10111147)
Rename all `Object` types to `RawObject` and remove the typedefs from steps 1
and 2. Add typedefs of the form `using Object = Handle<RawObject>;` and replace
`Handle<Object>` with `Object` for all types.

== Step 5+
Replace the `using Object = Handle<Object>;` typedefs with an actual class
hierarchy of Handles, moving member funtions from `RawObject` and friends over
to the new Handles. We should be able to convert a few types at a time, spread
across a number of diffs.

Based on Facebook D10106661